arch: isolate daemon contract from local lifecycle
