+++ src-tauri/crates/uc-daemon-client/Cargo.toml
@@ -11,9 +11,8 @@ path = "src/lib.rs"
 [dependencies]
 # Workspace crates
 uc-core = { path = "../uc-core" }
-uc-app = { path = "../uc-app" }
-uc-daemon = { path = "../uc-daemon" }
-uc-bootstrap = { path = "../uc-bootstrap" }
+uc-daemon-contract = { path = "../uc-daemon-contract" }
+uc-daemon-local = { path = "../uc-daemon-local" }
 
 # HTTP + WebSocket
 reqwest = { version = "0.12", default-features = false, features = ["json", "rustls-tls"] }
@@ -32,16 +31,10 @@ serde_json = "1"
 # URL parsing
 url = "2"
 
-# Tauri sidecar process management
-tauri-plugin-shell = { workspace = true }
-
 # Error handling + logging
 anyhow = "1.0"
 thiserror = "2.0"
 tracing = "0.1"
-
-[target.'cfg(unix)'.dependencies]
-libc = "0.2"
 
 [dev-dependencies]
 tokio = { version = "1", features = ["full"] }
