+++ src-tauri/crates/uc-daemon-contract/Cargo.toml
@@ -1,0 +1,18 @@
+[package]
+name = "uc-daemon-contract"
+version.workspace = true
+edition = "2021"
+description = "Daemon transport contracts shared by clients and server"
+
+[lib]
+name = "uc_daemon_contract"
+path = "src/lib.rs"
+
+[dependencies]
+uc-app = { path = "../uc-app" }
+uc-core = { path = "../uc-core" }
+base64 = "0.22"
+serde = { version = "1", features = ["derive"] }
+serde_json = "1"
+serde_with = "3.18.0"
+utoipa = { version = "4", features = ["uuid", "chrono", "url"] }
