+++ src-tauri/crates/uc-daemon-contract/src/api/dto/device.rs
