+++ src-tauri/crates/uc-daemon-contract/src/api/dto/encryption.rs
