+++ src-tauri/crates/uc-daemon-contract/src/api/dto/mod.rs
@@ -1,0 +1,7 @@
+pub mod clipboard;
+pub mod device;
+pub mod encryption;
+pub mod pairing;
+pub mod settings;
+pub mod setup;
+pub mod ws;
