+++ src-tauri/crates/uc-daemon-contract/src/api/mod.rs
@@ -1,0 +1,5 @@
+//! Shared daemon API contracts.
+
+pub mod auth;
+pub mod dto;
+pub mod types;
