+++ src-tauri/crates/uc-daemon-contract/src/lib.rs
@@ -1,0 +1,5 @@
+//! Shared daemon transport contracts.
+
+pub const DAEMON_API_REVISION: &str = "setup-pairing-http-routes-v1";
+
+pub mod api;
