+++ src-tauri/crates/uc-daemon-local/src/daemon_lifecycle.rs
