+++ src-tauri/crates/uc-daemon/src/api/dto/mod.rs
@@ -1,8 +1,4 @@
-pub mod clipboard;
-pub mod device;
-pub mod encryption;
 pub mod error;
-pub mod pairing;
-pub mod settings;
-pub mod setup;
-pub mod ws;
+pub use uc_daemon_contract::api::dto::{
+    clipboard, device, encryption, pairing, settings, setup, ws,
+};
