+++ src-tauri/crates/uc-daemon/src/lib.rs
@@ -6,7 +6,7 @@
 //! This crate is used as a library and as a binary (`uniclipboard-daemon`).
 
 pub const DAEMON_VERSION: &str = env!("CARGO_PKG_VERSION");
-pub const DAEMON_API_REVISION: &str = "setup-pairing-http-routes-v1";
+pub use uc_daemon_contract::DAEMON_API_REVISION;
 
 pub mod api;
 pub mod app;
