+++ src-tauri/crates/uc-tauri/src/commands/startup.rs
@@ -6,8 +6,8 @@ use std::sync::atomic::{AtomicBool, Ordering};
 use serde::Serialize;
 use tauri::AppHandle;
 use tracing::{info, info_span, Instrument};
-use uc_daemon::api::auth::DaemonConnectionInfo;
 use uc_daemon_client::DaemonConnectionState;
+use uc_daemon_contract::api::auth::DaemonConnectionInfo;
 use uc_platform::ports::observability::TraceMetadata;
 
 use crate::commands::record_trace_fields;
